requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Django==3.0.7
django-agora==0.4
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-seed==0.2.2
et-xmlfile==1.0.1
Faker==4.14.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
luhn==0.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
passwordgenerator==1.4
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
random-username==1.0.2
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
text-unidecode==1.3
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==1.2.0
XlsxWriter==1.3.7
